refactor(vECU): move server code out of core
